refactor: Update dependencies and providers
Update the dependencies in the package.json file to include the latest versions of "@apollo/client" and "@apollo/experimental-nextjs-app-support". Also, update the imports and context names in the page.jsx and layout.jsx files to use the correct casing for the ProgressContext. Finally, update the Providers component to include the ApolloNextAppProvider and configure the ApolloClient with the appropriate URI.